fix: resolve all ESLint errors and warnings
- Change let to const for capturedProgress array
- Remove unused _file parameter from isCompressionNeeded
- Extract React hooks to separate file to fix fast refresh warnings
- Wrap processVideoFrames in useCallback with proper dependencies
- Remove unused useContext import
