+++ src/contexts/UploadContext.tsx
@@ -1,4 +1,4 @@
-import { createContext, useContext, useState, useCallback, ReactNode } from 'react';
+import { createContext, useState, useCallback, ReactNode } from 'react';
 import { UploadProgress } from '../hooks/useUploadProgress';
 
 export interface BackgroundUpload {
@@ -21,7 +21,7 @@ interface UploadContextType {
   setActiveUpload: (recipeId: string | null) => void;
 }
 
-const UploadContext = createContext<UploadContextType | null>(null);
+export const UploadContext = createContext<UploadContextType | null>(null);
 
 export function UploadProvider({ children }: { children: ReactNode }) {
   const [backgroundUploads, setBackgroundUploads] = useState<BackgroundUpload[]>([]);
@@ -66,24 +66,3 @@ export function UploadProvider({ children }: { children: ReactNode }) {
   );
 }
 
-export function useUploadContext() {
-  const context = useContext(UploadContext);
-  if (!context) {
-    throw new Error('useUploadContext must be used within UploadProvider');
-  }
-  return context;
-}
-
-// Helper hook to get current active upload
-export function useActiveUpload(): BackgroundUpload | null {
-  const { backgroundUploads, activeUploadId } = useUploadContext();
-  return backgroundUploads.find(upload => upload.recipeId === activeUploadId) || null;
-}
-
-// Helper hook to get uploads in progress
-export function useUploadsInProgress(): BackgroundUpload[] {
-  const { backgroundUploads } = useUploadContext();
-  return backgroundUploads.filter(upload => 
-    upload.status !== 'completed' && upload.status !== 'failed'
-  );
-}
